Fix build errors for Data Overview
